+++ driver-core/CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 ------
 
 - [bug] Ignore static fields in mapper (JAVA-510)
+- [bug] Fix UDT parsing at init when using the default protocol version (JAVA-509)
 
 Merged from 2.0 branch: